Split out processLine
