+++ src/index.ts
@@ -12,6 +12,7 @@ import type {
 import * as cliProgress from 'cli-progress';
 import { downloadDumps } from './util/downloadDumps';
 import { readArgs } from './util/readArgs';
+import { processLine } from './yomitan/processLine';
 
 const outputZipName = (lang: string, date: string, version: string) =>
   `${lang} Wikipedia [${date}] (v${version}).zip`;
@@ -54,7 +55,9 @@ const outputZipName = (lang: string, date: string, version: string) =>
 
     while ((lineEnd = buffer.indexOf('\n')) !== -1) {
       const line = buffer.slice(0, lineEnd);
-      await processLine(line.trim(), dict, lang);
+
+      processLine(line.trim(), dict, lang);
+
       buffer = buffer.slice(lineEnd + 1);
       processedLines++;
 
@@ -86,85 +89,3 @@ div.gloss-sc-div[data-sc-wikipedia=term-specifier] {
 })().catch((e) => {
   console.error(e);
 });
-
-/**
- * Reads a line and adds that term entry to the dictionary
- */
-function processLine(line: string, dict: Dictionary, lang: string) {
-  const { term, termSlug, termSpecifier, reading, definition } = parseLine(
-    line,
-    lang
-  );
-
-  const termEntry = new TermEntry(term);
-  termEntry.setReading(reading);
-
-  const structuredContentList: StructuredContent[] = [];
-
-  // Add term specifier heading if exists
-  if (termSpecifier) {
-    const specifierSCNode: StructuredContentNode = {
-      tag: 'div',
-      content: `«${termSpecifier}»`,
-      data: {
-        wikipedia: 'term-specifier',
-      },
-      style: {
-        fontSize: '1.5em',
-      },
-    };
-
-    structuredContentList.push(specifierSCNode);
-  }
-
-  const definitionStrings = definition.split('\\n').map((line) => line.trim());
-  const definitionUList: StructuredContentNode = {
-    tag: 'ul',
-    content: definitionStrings.map((definitionString) => ({
-      tag: 'li',
-      content: definitionString,
-    })),
-    data: {
-      wikipedia: 'abstract',
-    },
-  };
-  structuredContentList.push(definitionUList);
-
-  // Read more
-  const articleLink = `https://${lang.toLowerCase()}.wikipedia.org/wiki/${termSlug}`;
-  const readTheRest =
-    lang === languages.ja
-      ? '続きを読む'
-      : lang === languages.zh
-      ? '查看更多'
-      : 'Read more';
-  const linkSC: StructuredContentNode = {
-    tag: 'ul',
-    content: [
-      {
-        tag: 'li',
-        content: [
-          {
-            tag: 'a',
-            href: articleLink,
-            content: readTheRest,
-          },
-        ],
-      },
-    ],
-    data: {
-      wikipedia: 'continue-reading',
-    },
-    style: {
-      listStyleType: `"${linkCharacter}"`,
-    },
-  };
-  structuredContentList.push(linkSC);
-
-  termEntry.addDetailedDefinition({
-    type: 'structured-content',
-    content: structuredContentList,
-  });
-
-  dict.addTerm(termEntry.build());
-}
